Authentication and authorization done

+++ notes-app-back-end/src/api/notes/handler.js
@@ -16,8 +16,9 @@ class NotesHandler{
         this._validator.validateNotePayload(request.payload);
 
         const { title = 'untitled', body, tags } = request.payload;
-
-        const noteId = await this._service.addNote({title, body, tags});
+        const { id: credentialId } = request.auth.credentials
+        
+        const noteId = await this._service.addNote({title, body, tags, owner:credentialId});
         const response = h.response({
             status: 'success',
             message: 'Catatan berhasil ditambahkan',
@@ -53,8 +54,9 @@ class NotesHandler{
         // }
     }
 
-    async getNotesHandler(){
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request){
+        const { id: credentialId } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
         return{
             status: 'success',
             data: {
@@ -65,13 +67,16 @@ class NotesHandler{
 
     async getNoteByIdHandler(request, h){
         const { id } = request.params;
-            const note = await this._service.getNoteById(id);
-            return {
-                status: 'success',
-                data: {
-                  note,
-                },
-              };
+        const { id: credentialId } = request.auth.credentials
+
+        await this._service.verifyNoteOwner(id, credentialId)
+        const note = await this._service.getNoteById(id);
+        return {
+            status: 'success',
+            data: {
+                note,
+            },
+            };
         
         // try {
             
@@ -95,11 +100,13 @@ class NotesHandler{
         // }
     }
 
-    putNoteByIdHandler(request, h){
+    async putNoteByIdHandler(request, h){
         this._validator.validateNotePayload(request.payload);
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
 
-        this._service.editNoteById(id, request.payload);
+        await this._service.verifyNoteOwner(id, credentialId)
+        await this._service.editNoteById(id, request.payload);
             return{
                 status: 'success',
                 message: 'Catatan berhasil diperbarui'
@@ -128,9 +135,12 @@ class NotesHandler{
         // }
     }
 
-    deleteNoteByIdHandler(request, h){
+    async deleteNoteByIdHandler(request, h){
         const { id } = request.params;
-        this._service.deleteNoteById(id);
+        const { id: credentialId } = request.auth.credentials
+
+        await this._service.verifyNoteOwner(id, credentialId)
+        await this._service.deleteNoteById(id);
             return{
                 status: 'success',
                 message: 'Catatan berhasil dihapus'
